Moved Map to separate component
