+++ src/components/Map/Map.scss
@@ -1,0 +1,4 @@
+.map-container {
+    width: 100%;
+    height: 100%;
+}
