--- src/pages/Order/Order.scss
@@ -1,4 +1,0 @@
-.map-container {
-    width: 100%;
-    height: 100%;
-}
